Moved ExampleContract into its own crate
