+++ Cargo.toml
@@ -3,9 +3,10 @@ resolver = "2" # As of Nov 2021, rust doesn't seem to use new resolver if you do
 
 members = [
     "client/codegen_wrapped",
+    "client/contract_example",
+    "client/contract_macro",
     "client/rust",
     "client/serialize/rust",
-    "client/contract_macro",
     "crates/better_borrow",
     "crates/datastructures/arena",
     "crates/rsharp/codegen",
