+++ client/rust/src/baseline.rs
@@ -379,7 +379,6 @@ pub mod c_api {
 
     use super::*;
     use crate::contract::c_api::ContractDataHandle as CContractDataHandle;
-    use crate::contract::c_api::ExampleContract;
     use crate::contract::properties::primitives;
     use crate::object::c_api::ObjectHandle as CObjectHandle;
 
@@ -519,17 +518,4 @@ pub mod c_api {
 
     // This is like doing `monomorphize!("whatever", Keyframe, u8, u16, ...)
     primitives!(; types, monomorphize, "tp_client::baseline");
-
-    // ---- Example contract for tests ---- //
-
-    #[remangle(substitute!())]
-    #[ffi_export]
-    pub fn Baseline__register_contract___example<'a>(
-        baseline: &'a mut Baseline,
-    ) -> repr_c::Box<ExampleContract> {
-        let c: ExampleContract = baseline
-            .register_contract()
-            .expect("Failed to register contract");
-        Box::new(c).into()
-    }
 }
