+++ client/rust/src/contract/mod.rs
@@ -171,179 +171,4 @@ pub mod c_api {
     pub fn ContractId__version<'a>(cid: &'a ContractId) -> ContractId_Version {
         cid.version.into()
     }
-
-    mod _ExampleStates {
-        use crate::contract::properties::states::id::c_api::{
-            StateId_F32, StateId_I8, StateId_String, StateId_U8,
-        };
-        use crate::contract::states;
-
-        use rsharp::remangle;
-        use safer_ffi::prelude::*;
-
-        #[remangle(substitute!())]
-        #[states]
-        #[derive_ReprC]
-        #[ReprC::opaque]
-        #[derive(Clone)]
-        pub struct ExampleStates {
-            u8_0: u8,
-            u8_1: u8,
-            i8_0: i8,
-            i8_1: i8,
-            f32_0: f32,
-            f32_1: f32,
-            str_0: String,
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__u8_0(s: &ExampleStates) -> repr_c::Box<StateId_U8> {
-            Box::new(StateId_U8::from(s.u8_0)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__u8_1(s: &ExampleStates) -> repr_c::Box<StateId_U8> {
-            Box::new(StateId_U8::from(s.u8_1)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__i8_0(s: &ExampleStates) -> repr_c::Box<StateId_I8> {
-            Box::new(StateId_I8::from(s.i8_0)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__i8_1(s: &ExampleStates) -> repr_c::Box<StateId_I8> {
-            Box::new(StateId_I8::from(s.i8_1)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__f32_0(s: &ExampleStates) -> repr_c::Box<StateId_F32> {
-            Box::new(StateId_F32::from(s.f32_0)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__f32_1(s: &ExampleStates) -> repr_c::Box<StateId_F32> {
-            Box::new(StateId_F32::from(s.f32_1)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleStates__str_0(s: &ExampleStates) -> repr_c::Box<StateId_String> {
-            Box::new(StateId_String::from(s.str_0)).into()
-        }
-    }
-    pub use _ExampleStates::ExampleStates;
-
-    mod _ExampleContract {
-        use super::ContractDataHandle as CContractDataHandle;
-        use super::ExampleStates;
-        use crate::baseline::Baseline;
-        use crate::contract::properties::dynamic::DynTpProperty;
-        use crate::contract::{Contract, ContractDataHandle, ContractId};
-        use crate::object::c_api::ObjectHandle as CObjectHandle;
-
-        use rsharp::remangle;
-        use safer_ffi::prelude::*;
-
-        #[remangle(substitute!())]
-        #[derive_ReprC]
-        #[ReprC::opaque]
-        pub struct ExampleContract {
-            handle: ContractDataHandle,
-            states: ExampleStates,
-            channels: (),
-        }
-        impl Contract for ExampleContract {
-            type States = ExampleStates;
-
-            type Channels = ();
-
-            const ID: ContractId = ContractId {
-                name: "teleportal.example-ffi-contract",
-                version: (1, 2, 3),
-            };
-
-            fn new(handle: ContractDataHandle) -> Self {
-                Self {
-                    handle,
-                    states: ExampleStates::new(handle),
-                    channels: (),
-                }
-            }
-
-            fn states(&self) -> &Self::States {
-                &self.states
-            }
-
-            fn channels(&self) -> &Self::Channels {
-                &self.channels
-            }
-
-            fn handle(&self) -> ContractDataHandle {
-                self.handle
-            }
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleContract__drop(c: repr_c::Box<ExampleContract>) {
-            drop(c)
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleContract__handle(c: &ExampleContract) -> repr_c::Box<CContractDataHandle> {
-            Box::new(CContractDataHandle::from(c.handle())).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleContract__states<'a>(c: &'a ExampleContract) -> &'a ExampleStates {
-            c.states()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleContract__object_create(
-            contract: &ExampleContract,
-            baseline: &mut Baseline,
-            u8_0: u8,
-            u8_1: u8,
-            i8_0: i8,
-            i8_1: i8,
-            f32_0: f32,
-            f32_1: f32,
-        ) -> repr_c::Box<CObjectHandle> {
-            let states = [
-                DynTpProperty::Primitive(u8_0.into()),
-                DynTpProperty::Primitive(u8_1.into()),
-                DynTpProperty::Primitive(i8_0.into()),
-                DynTpProperty::Primitive(i8_1.into()),
-                DynTpProperty::Primitive(f32_0.into()),
-                DynTpProperty::Primitive(f32_1.into()),
-            ];
-            let obj = baseline
-                .object_create(contract, states.into_iter(), [].into_iter())
-                .unwrap();
-            Box::new(CObjectHandle::from(obj)).into()
-        }
-
-        #[remangle(substitute!())]
-        #[ffi_export]
-        pub fn ExampleContract__object_remove(
-            baseline: &mut Baseline,
-            obj: repr_c::Box<CObjectHandle>,
-        ) {
-            baseline
-                .object_remove::<ExampleContract>(obj.inner)
-                .unwrap()
-        }
-    }
-    pub use _ExampleContract::ExampleContract;
 }
