+++ client/serialize/rust/Cargo.toml
@@ -16,5 +16,8 @@ bimap = "0.6"
 paste = "1"
 dynpath = { git = "https://github.com/TheButlah/dynpath", rev = "0058f9d5fd28cc9760f8c4cbb99974e060297dde" }
 
+[dev-dependencies]
+tp_contract_example = {path = "../../contract_example"}
+
 [build-dependencies]
 flatc-rust = "0.2"
