+++ crates/rsharp/rust/src/string.rs
@@ -1,6 +1,7 @@
 #![allow(non_snake_case)]
 
 use crate::remangle;
+
 use derive_more::{From, Into};
 use ref_cast::RefCast;
 use safer_ffi::prelude::*;
